Chatbot/requirements.txt
# for running : pip install -r requirements.txt

# Web framework for APIs
fastapi==0.109.0
uvicorn==0.27.0

# Database
sqlalchemy==2.0.25

# Excel handling
pandas==2.1.4
openpyxl==3.1.2

# Langchain for RAG system
langchain==0.1.7
langchain-openai==0.0.7

# Vector databases
chromadb==0.4.22  # If using ChromaDB

# Embeddings & LLMs
openai==1.11.1
sentence-transformers==2.2.2
transformers==4.30.2  # For tokenization and model handling

# Streamlit for UI
streamlit==1.30.0

# Other utilities
python-dotenv==1.0.1  
requests==2.31.0  
torch==2.1.0  
google-generativeai==0.1.0  
